+++ NEWS.md
@@ -1,3 +1,14 @@
+# motif 0.6.1
+
+* `lsp_add_stars` and `lsp_add_terra` now do not drop the `signature` list-column, but restructures it into many raster layers
+* Changes defaults in `lsp_signature`, `lsp_search`, and `lsp_compare` to `ordered = FALSE` and `repeated = FALSE` (both were `TRUE` before)
+* Improves handling of terra's `SpatRaster` class in `lsp_extract`
+* Mutes `philentropy::distance` messages
+
+# motif 0.6.0
+
+* Adds possibility of providing two numerical values to `window`, where the first one represents the expected window size and the second one is the window shift
+
 # motif 0.5.2
 
 * Speeds up distance calculations in `lsp_search` and `lsp_compare`